defined tnfsPacket_t type
